refactor: remove redundant package from mapper import
