+++ remote_datasource/src/main/java/com/baghdad/remoteDataSource/RemoteGenreDataSourceImpl.kt
@@ -1,6 +1,6 @@
 package com.baghdad.remoteDataSource
 
-import com.baghdad.remoteDataSource.mapper.movie.toDto
+import com.baghdad.remoteDataSource.mapper.toDto
 import com.baghdad.remoteDataSource.response.GenreListResponse
 import com.baghdad.remoteDataSource.util.handleRequest
 import com.baghdad.repository.datasource.remote.RemoteGenreDataSource
